requirements.txt
openai
python-dotenv
pandas
fpdf